+++ core-bundle/src/Resources/contao/modules/ModuleCustomnav.php
@@ -129,8 +129,8 @@ class ModuleCustomnav extends \Module
 		{
 			$_groups = \StringUtil::deserialize($objModel->groups);
 
-			// Do not show protected pages unless a back end or front end user is logged in
-			if (!$objModel->protected || BE_USER_LOGGED_IN || (is_array($_groups) && count(array_intersect($_groups, $groups))) || $this->showProtected)
+			// Do not show protected pages unless a front end user is logged in
+			if (!$objModel->protected || (is_array($_groups) && count(array_intersect($_groups, $groups))) || $this->showProtected)
 			{
 				// Get href
 				switch ($objModel->type)