+++ core-bundle/src/Resources/contao/modules/ModuleQuicknav.php
@@ -147,8 +147,8 @@ class ModuleQuicknav extends \Module
 				$objSubpage->domain = $host;
 			}
 
-			// Do not show protected pages unless a back end or front end user is logged in
-			if (!$objSubpage->protected || (!is_array($_groups) && FE_USER_LOGGED_IN) || BE_USER_LOGGED_IN || (is_array($_groups) && array_intersect($_groups, $groups)) || $this->showProtected)
+			// Do not show protected pages unless a front end user is logged in
+			if (!$objSubpage->protected || (is_array($_groups) && array_intersect($_groups, $groups)) || $this->showProtected)
 			{
 				// Do not skip the current page here! (see #4523)