+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.3.0-beta1 (2016-XX-XX)
 
+ * Do not show protected elements if "show published elements" is enabled (see contao/core#8149).
  * Scroll form error messages into view (see contao/core#8044).
  * Add a "title" field to the meta wizard (see #601).
  * Support using namespaces and use statements in DCA and config files (see contao/core#8530).